Search backend: convert closures and channels to stream (#37509)
This modifies the stream-to-http code to fit into our normal streaming.Sender interface.

We jumped through a lot of hoops to convert a callback-based stream into a set of event channels, and it was honestly very confusing. This rewrites the code that handles converting events, populating metadata, and periodically flushing the results to the underlying stream into a type that implements streaming.Sender. This allows us to pass it directly into search.Execute and gets rid of a bunch of closures that implicitly capture a large amount of state.