process results in parallel
This updates our event handler to process results in parallel. It uses
`group.ParallelOrdered`, which maintains the order the events were
streamed in, but still allows us to do things like fetch repo metadata
in parallel.

Repo metadata is not particularly expensive, so this isn't a big win
yet, but it unlocks more expensive decorations like streamed syntax
highlighting.