policies: update matched repos right after policy is created
This is a short/medium-term solution to the problem that it can take up
to 1min for policies to have an effect after creating/update.

Change is motivated by me trying to write integration for
lockfile-indexing-based dependency search and the tests have to wait for
the lockfile policy to take effect.

Without the change in this PR this takes ~1min, because the matcher
only runs every minute and checks whether repos match a new policy.

With this change, we match the repositories right after creating policy.