return early when prefiltering returns no results
This adds an early return to `filteredStructuralSearch` when we return
no matches from `regexSearchBatch`. Before this, we would continue to do
a structural search on the repo with an empty string, which is
apparently interpreted by Comby as "search everything", which was
causing unindexed searches to take far more resources than they needed to.