codemirror: Take focus state into account for filter highlighting on every update
Currently when navigating e.g. from search results to a file, we will
update the content of the search query input but not focus it. Because
the content changes the logic for recomputing the current filter
highlight gets triggered and highlights the last filter in the input,
which is not desirable.

This commit updates the logic to take the focus state into account on
every update.