batches: make current execution "clickable" when using VoiceOver
Executions that aren't current operate as expected, but the tree for
current specs looks something like this, and VoiceOver doesn't make it
activatable:

```
H3
|
+-- Icon
+-- Link
```

Because the element is a heading and not a group, the user can't
navigate into the child elements. We also can't ignore the icon entirely
because it includes important semantic information (namely, that this is
the current spec).

There's no real reason we can't flip this so the icon is inside the
link, so that's what I've done. There's a very slight visual regression
on hover where the link underline now extends into the space between the
icon and text, but I think that's fine, and the icon being clickable
really feels right anyway, so it's visually part of the label anyway.

Related to #34084.