Fix scrolling first selected line into view
Problem: When selecting multiple lines such that the first selected line
is out of view, the file view jumps back to the first selected line
after selecting the last line.

This is because we are only considering the first selected line when we
determine the position to scroll to.

This commit changes the logic to take into account the top and bottom of
the complete line range.