Simplify blob view line highlighting logic
Currently the line highlighting logic is triggered on every re-render of
the component because the `nextCodeView` observable emits unconditionally
on every render.

This was introduced to ensure that the highlighting logic runs after
React updated the DOM, otherwise React would overwrite our changes.

In that case a simpler solution that avoids updating the highlighted lines
if neither code nor position is to avoid using observables altogether and
put the highlighting logic directly in `useEffect`, conditioned on file and
position changes.