Fix duplicate history entry on selecting lines (#38193)
It appears that we are adding the exact same history twice, making the
back and forward buttons behave unexpectedly.

Both the function this line was delete from as well as the observable
that gets triggered by calling `nextPopOverClose` add a history entry. I believe
calling `nextPopOverClose` is unncessary here (or we leave it but remove
the call to `history.push(...)` below).