Avoid duplicate history entries for the same line position (#38204)
In a way a continuation of #38193.

In noticed that clicking the same line multiple times adds a separate
history entry for every click. That's not a great experience wrt
navigating back and forth.

This is done by comparing the URL query parameters and their values.
Location is encoded as a query parameter without a value, so if the
location changed then trying to look up the same paremeter in the new
URL will return `null` (and therefore be different from an empty
string).