codemirror: Fix suggestions snippet behavior (#38257)
Tabbing from the last field in a snippet causes the input to loose
focus. CodeMirror only overwrites the Tab key behavior for any but the
last placeholder field in the snippet.
The simple solution is to always append an empty placeholder field at
the end of the snippet. This replicates Monaco's behavior.