Search backend: add PartialJob (#38321)
This adds PartialJob to the internal/search/job package. The goal of
this type is to represent partially constructed jobs by the type system.
Right now, the only time we do something like this is with the repo
pager job, but the goal is to use this when converting more predicates
to just be jobs.