search: Show token information for operators (CodeMirror) (#38336)
This commit extends the token info extension to
highlight and show tooltip information for AND and OR operators.

To do this I updated the editor state to also compute a parse tree from
the current query tokens. If an operator is hovered I'm locating the
corresponding node in the parse tree and use that nodes range
information to highlight the operator and its operands.

If the query cannot be parsed as a parse tree it falls back to just
highlighting the operator.