search: Add placeholder to fallback query input
We are rendering a simple text input until CodeMirror/Monaco are loaded.
The current experience with CodeMirror is not ideal because we show an
empty input and as soon as CodeMirror loads it shows a placeholder text.

This commit adds logic to show a placeholder text in the plaintext input
if CodeMirror is selected. There might still be a noticable difference
between the browser's builtin placeholder color and CodeMirror's
placeholder color, but that's still much better than showing no
placeholder at all.

I'm also adding back setting the initial editor value immediately at
creation time (instead of also using a transaction to set the
initial value). Otherwise a flash of content can happen on the result
page when we first load the fallback input (with value), then load
CodeMirror (without value) and then set the value (via transaction).

I'm changing the font size for the fallback input to the same one we use
for CodeMirror. It already doesn't match the font size for Monaco anyway
(I see a noticable difference in font size in both Firefox and Chrome),
so it doesn't make things worse.