Search backend: unify job self-descriptions (#38353)
This PR unifies the concept of a self-describing job around the Describer interface .Name and Fields (formerly Tags) both already existed, but Children() is new. With these three methods, we can implement all the pretty printers we currently use (JSON, Sexp, Mermaid) without them needing to unwrap the interface. This means we can move the pretty-printing code out of the catch-all job package that necessarily lists every possible job. This is a big deal because, with the current setup, any dependency on the pretty printers means a dependency on every package that defines a job, which is a big pain because of circular dependencies.