authz: Allow directory read access for sub repo permissions
If a path is allowed by sub repo permissions, we should also allow read
access to any directory, but only the directory, that live above the
item.

For exaple, if we allow access to "foo/bar/data.txt" we should also
allow the path "foo/" and "foo/bar/" so that we can navigate down the
tree.

This does NOT imply access to all items living under "foo/" or
"foo/bar/"