JetBrains: Fix encoding errors when using the standalone version (#38474)
This PR fixes a text encoding error with the standalone version. Not
urgent but this was annoying me 😐.

The issue is that the default `atob` method does not properly handle
unicode characters. We can, however, use the same library that we use to
encode the text. This also preserves leading [ZERO WIDTH NO-BREAK
SPACE](https://www.fileformat.info/info/unicode/char/feff/index.htm)
character, a property that the `TextEncoder` did not help with.