Re-enable flaky dependency search test
This test was disabled in #38505 because the dependency search for
`sgtest/poetry-hw` didn't return `atomicwrites@1.4.0`. See these two
failures:

- https://buildkite.com/sourcegraph/sourcegraph/builds/159353#0181df60-0721-4b24-aa3d-8247cc21d431/278-279
- https://buildkite.com/sourcegraph/sourcegraph/builds/159359#0181df7d-3e35-433c-bd78-b758dbde3bdd/278-280

The failures don't look that flaky to me:

- why is it `atomicwrites` that's missing, every time?
- why could it be cloned (which we assert before in the test) but the rev wasn't returned in response?

I tried to reproduce it locally but couldn't. Multiple runs (with fresh
instance, warm instance) and no failure. I re-ran it in CI [20 times](https://buildkite.com/sourcegraph/sourcegraph/builds/159516) and
it didn't fail.

That made me curious about `atomicwrites@1.4.0` so I looked at its pypi
releases: https://pypi.org/project/atomicwrites/#history

Turns out a new version - 1.4.1 - was released on July 8, the same day
that these failures started to show up. And the older releases now have
July 9 (!) attached to them.

The maintainer also set the package to "unmaintained" on the same day:

https://github.com/untitaker/python-atomicwrites/commit/d18328460520e18b4f197297f962d4444c5889b6

What to make of this? Not sure. But I somehow suspect that the release
process for atomicwrites@1.4.1 got a bit mixed up on July 8 and maybe
1.4.0 was un-clone-able for some time?

Since I can't reproduce it locally and it passed 20+ times in CI, I
think it's fine to re-enable this and mark it as a "glitch".