search: Show placeholder for some filters (CodeMirror)
This commit adds logic to render a simple placeholder when a filter has
no value. I only added placeholders to filters for which we don't
provided static suggestions. I think suggestions provide enough
information about which value can be used. This way we also avoid some
visual clutter (placeholder + suggestions visible otherwise).

Note that the filter will only be shown when the filter is currently
"active", i.e. the text cursor is "inside" the filter (and there is no
value of course).

We could also make it so that the placeholder is always shown,
regardless where the cursor is positioned.