Stop sending GET requests to test if a package exists
We already send a GET request to download the package, which fails if
the pakage doesn't exist. Before this change, we always sent a GET
request for every tracked version (2.3m total versions for crates.io) every
time gitserver synced package repos. After this change, we only send
GET requests for packages that are missing locally.

There is a risk that the contents of a package changes in case the
package host overwrites an older release. Currently, the only way
Sourcegraph can pick up the new contents is by manually removing the
package repo from gitserver. This commit doesn't change anything here
because this is true both before this commit and after this commit.