Deprecate enableGitServerCommandExecFilter and make it NOOP (#38642)
Setting this flag will now have no effect on the code and the CVE fix
it was intended for will now be part of the core behaviour and cannot
be overridden.

In this commit we remove all references to
EnableGitServerCommandExecFilter in the tests and remove the tests for
the old behaviour as they reflect the incorrect behaviour now. The
tests that were updated reflect the correct behaviour - which is to
return an 400 HTTP response instead.