Fix regressions in the Python and npm package host (#38652)
* Fix regression in Python package host

Previously, the code tried to download from an empty URL because
`PythonVersionedPackage.PackageURL` could be an empty string. Now, the
field has been removed and we find the `PackageURL` inside the
`Download()` function.

* Fix regression for npm package host

Similar to the Python package host, the TarballURL was empty because we
no longer called the `Get()` function before invoking the `Download()`
function.