search: Fix highlight filter focus bug
It's unsatisfying finding a solution to a problem and not understanding
why the solution fixes the problem... this is the case here. I don't
know how the syntax highlighting extension could interfere with the
other extension or the editor focus, especially since it's only executed
when the query changes.

But alas, moving the extension after the current filter highlighting
extension fixes the issue. ¯\_(ツ)_/¯