remove `repo:has.description`-specific completion code
Whether this code is present or not doesn't seem to make a difference in behavior, or tests, so removing it. Some, but not all, `repo:has.description`-specific completion code was removed in 6f2d13a5, which caused predicates to be the _only_ suggestions visible after typing in `repo:`. Completely removing this code appears to preserve the original behavior.