zoekt: use GOGC=25
After much experimentation we have a new value for GOGC, which is 25. We
noticed a large regression in zoekts memory use when we switched to
go1.18. So we reverted to go1.17 in zoekt releases and since then have
been experimenting with different GOGC values. We have found the same
stable heap size with GOGC=25 and have had this stable size for more
than a week.

Included zoekt changes:

- 16b6b7e50c all: remove datadog support
- f7f68fd7f3 ci: cleanup whitespace
- 42276b7b70 docker: change our default GOGC to 25

Test Plan: Lots of experimentation in our production work loads has been
done. This is the outcome of that.