docs: move package host admin docs to other code host docs (#38745)
This is the 2nd step of what I started yesterday: it splits the
docs into two. Into "integration" docs and "how to setup a code host connection".

That mirrors what we've done for the other code hosts. It's not perfect
and I think there's still a lot of other improvements to be made, but I
think consistency is already helpful here.