Code Insights: Unbind creation and edit UI pages and insight models  (#38739)
Prior to this PR, we had creation and edit UI forms in which we had explicit binding between the insight model and its permissions and UI itself. This was a problem because we couldn't freely reuse UI across different creation UI forms. This PR bubbles up all common parts and unbinds permissions calls from the creation UI forms.

In this PR we unify form UI that this UI doesn't know anything about its possible consumers
and lift all dependent logic to the specific page level (creation or edit UI pages)