zoekt: directly use GitHub API for update script (#38609)
This greatly speeds up the script since it avoids the use of using
GOPROXY=direct. It also allows us to inline the suggested PR and commit
descriptions. The intention here is we can then start using this output
to improve the automated PRs created by the bot.

Additionally I changed the command we test to be an package which is
faster to compile.

This is some over engineering thanks to me having only 1h before a
chore.

Included zoekt changes:

- 16b6b7e50c all: remove datadog support
- f7f68fd7f3 ci: cleanup whitespace

Test Plan: ran the script and included the update to gomod here.