search: Use @codemirror/lint and add support for actions (#38484)
This commit switches our CodeMirror diagnostics implementation from a
custom extension to the one provided by CodeMirror. It provides
diagnostic markers and the hovers show a colored bar (matching severity
type) next to the text.

I also took the opportunity to refactor our diagnostics implementation
to be less Monaco specific. Instead there is now an intermediate layer
(which is inspired by CodeMirror) and the editors themselves convert
those objects to something appropriate for them.

I also added support for actions, which can be seen in the video. We can
change the default style of the buttons if necessary.