Allow package hosts to sync versions on-demand
Previously, the `/compare` page failed to load the diff if
one of the versions was not tracked in our database. This commit fixes
that issue by allowing the package host to lazily sync the requested
versions (for the diff). The fix is generic so this lazy syncing logic
works in other pages like the blob view if the user manually types the
revision in the URL.