blob: Add experimental support for rendering files with CodeMirror
This commit introduces a file view backed by CodeMirror. It shows the
file as plain text without syntax highlighting.
I spent a little bit of time to implement a line gutter that supports
line (range) selection.

To use the CodeMirror version
`experimentalFeatures.enableCodeMirrorFileView` needs to be set to
`true`.

Note that this is a rough minimal drop in replacement and the way things
are implemented might not be optimal.