cm blob view: Improve initial render; scroll into view and range selection ergonomics
This commit addresses minor issues with the initial implementation:

The `useExtension` hook as it were resulted in creating the editor
without the configured extensions and only adding them via a transaction
afterwards. While that should generally be fine, if something like dark
theme is set via such an extension (as it is in our case), the incorrect
theme might be briefly visible.
To fix this the hook was completely redesigned: It's now officially a
compartment hook. The extension(s) now has to be created inside the
component and passed to the hook. The hook will return a compartment
extension and a function to update said compartment.
This allows the component to compute the extensions via `useMemo` and
update the compartment via `useEffect`.

When a line range was selected by first clicking the start line and then
shift-clicking the end line, two history entries would be added (one for
the first click and one for the second click).
The `selectableLineNumbers` extension now exposes the `MouseEvent` to
the `onSelection` with which the handler can decide what to do on
shift+click. In this specific instance the handler will replace the
current entry instead of adding a new one. For the most common case
(click start line + click end line) this will create a more pleasant
experience.

The "scroll into view" logic was a bit inconsistent: Because it wasn't
clear which lines are actually _visible_, the effect would be configured
with `"nearest"` to ensure that the line was definitely visible. But
that leads to an inconsistent behavior.
While I'm not sure whether that's the current way to do it, it's
actually possible to determine whether or not a line is visible by
getting its vertical position and compare it against the scroll
containers scroll position.
Now a selected line is always centered when it is not visible (whether
or not _that_ behavior is the best UX is a different question)