search: Show Notepad only on certain pages
Closes #38521

This commit changes on which pages the notepad is shown. Now the notepad
will be shown on the search results page, the notebooks list page (so
that there is a visiual indicating that somethin happens) and any
repo/file related pages (that would also include pages that don't show
code (e.g. branches) but I think it's an OK tradeoff)

In order to have easier access to route related information I moved the
notepad to `Layout`.

I don't think this needs to be put behind a feature flag, but if you
think otherwise, please let me know.