worker: use conf instead of globals to read PermissionsUserMapping
Globals depends on initialization which we haven't done in worker. We
could just do the initialization, but it feels like the next change we
should just remove globals and rely on directly speaking to conf. But
will discuss that soon.

As such we always used the default value for PermissionsUserMapping
inside of worker, which is email. So using username would fail. Now it
works.

Test Plan: Ran setRepositoryPermissionsForBitbucketProject manually
against the devserver with bindid set to username. Before this change
this call would not result in repositories becoming visible for the
user. After the change the repositories are visible.