worker: start WatchPermissionsUserMapping at startup (#39035)
Globals depends on initialization which we haven't done in worker. This lead
us to always using the default value in worker. As such we always used the
default value for PermissionsUserMapping inside of worker, which is email. So
using username would fail. Now it works.

Test Plan: Ran setRepositoryPermissionsForBitbucketProject manually against
the devserver with bindid set to username. Before this change this call would
not result in repositories becoming visible for the user. After the change the
repositories are visible.