Support syntax highlighting in CodeMirror file view
This commit adds support for syntax highlighting from lsif/scip data that is
returned for some languages. Updating the backend to return lsif/scip
data for every language is done in a separate PR.

Syntax highlighting is done by an extension that takes the JSON encoded
scip data and converts it to something CodeMirror can understand.
Decorations are only generated for the lines that are currently
rendered. I'm using binary search to find the corresponding data in the
highlighting data (whether or not it makes to lazily generate the
decorations is something we can still evaluate in the future).

Other auxiliary changes in this commit:

- Changed the base CodeMirror hook to allow for "manually" dispatching a
  transaction to update the value. Without this we would trigger at
  least two transactions when loading a file: One for updating the
  document and one for updating the syntax highlighting.
  Now we can do both in a single transaction.
- Added a folder for CodeMirror blob extensions